A shape should support unknown fields when declared with '...' even if 'promote_nullable_to_optional_in_shapes' is disabled
Summary:
The intention with the `promote_nullable_to_optional_in_shapes` experiment was, when enabled, to modify the interpretation of any shape such that:

  1. Each of its nullable fields are considered optional as well
  2. The shape always supports unknown fields

I made a mistake when I originally made this experiment, and accidentally set `si_allows_unknown_fields` to be **equal** to the experiment's value. This means, if I were to disable the experiment, that shapes explicitly declared as supporting unknown fields (e.g., `shape(...)`) would not be interpreted as such.

Reviewed By: dlreeves

Differential Revision: D5412505

fbshipit-source-id: 239d9ba6d65987cbf2ed19df60773ae8130644a8