Reduce race condition when smashing on PPC64
Currently on PPC64 architecture all immediates are loaded using up to 5
instructions, which makes the read not atomic.
The smashable instructions may have its immediate changed on the fly, i.e.,
while running or re-emitting the JITted code, which causes race condition,
if reading the immediate is composed with more than one instruction.
The race condition happen at two different moments:

1. The JITted code is being executed and the pipeline stop reading the
instructions group that reads the immediate in the middle (thread
preemption), then another thread smashes the instruction that uses the same
immediate.

2. The Decoder is reading the immediate used in a smashable instruction and
this thread stop being executed (thread preemption), then another thread
smashes the instruction that uses the same immediate.

The item 2 represents the most frequently cause of the race conditions, as
the Decoder code takes much more instructions (written in C++) to read the
entire immediate, when comparing the reading while executing (maximum
of 5 instructions).
For the reason above, this PR is being sent to fix the item 2, which locks
the C++ methods that read/write smachable instructions.
The lock was not extended while executing the JITted code, that would fix
the item 1, because it would create a performance bottleneck. Fixing it will
need a bigger change, as many parts of the code depends on how the immediate
is being written/loaded currently, and will be sent later, after be better
analyzed.