Moving context building code out of keyword autocomplete.
Summary:
I'd like to use the context for other types of completions as well for determining whether or not a certain type of completion should be done at a given location, so I'm pulling this out into its own file.

This is also a set-up task for the next task in this stack where I'm handling autocomplete in empty locations.

Reviewed By: dabek

Differential Revision: D5402434

fbshipit-source-id: 35bc435f889640978d2615e530b47ca2495573a6