add logging of jsonrpc responses as well
Summary:
We were already logging the jsonrpc messages that we receive.

This diff also logs the jsonrpc responses that we send back (in the case of incoming requests).

In future I'd like to do cleaner logging for other cases... (1) when we issue requests to the LSP editor, and subsequently get a response; (2) when we receive a notification from hh_server, and so send a jsonrpc notification to the editor. But I'll leave that for future.

Reviewed By: arxanas

Differential Revision: D5343183

fbshipit-source-id: 3b225053ec6aeacb017313930cb2de7eaf1dac38