initial cut at tool for sending batch commands to the LSP.
Summary:
This tool is the beginning of a set of work to better integration test the LSP.  The capabilities of this tool
will ultimately get into test/integration for hack, but this makes it possible for anybody to send arbitrary
commands to the LSP and see how it responses.  Should come in handy during debugging.

Reviewed By: arxanas

Differential Revision: D5406708

fbshipit-source-id: c0b72af2168813a2dad2839cc974939608326040