Add 'kind' field to autocomplete results
Summary:
For LSP purposes we want to record the 'kind' of each autocomplete result - is it a method, a class, an interface, a variable, ...? These kinds will be represented with icons in the autocomplete menu.

(Currently Hack only has two kinds, "function/constructor" which it represents with the letter f, and "everything else" which it represents with the absence of an icon).

(I think icons are really valuable because they distinguish which autocomplete results come from a real language service, vs which autocomplete results come just from Atom's ctags or similar).

Reviewed By: dabek

Differential Revision: D5366150

fbshipit-source-id: f5a1e8de3e1c7cd9a1813ad2ac83b6caaa5458ab