Add option for autocomplete to delimit on namespaces
Summary:
The current behavior of autocomplete doesn't recognize namespace characters as anything special. It just says that a method identifier or class identifier might include `\`, much like it includes `_` or any other character. It leaves it to the downstream caller to figure out what it wants to do with namespaces. The autocomplete results look like this:
  St|    =>  Str\contains (method)
  Str\c  =>  Str\contains (method)

This diff introduces a new mode of autocomplete, one which does recognize namespaces as first-class entities:
  St|    =>  Str (namespace)
  Str\c  =>  contains (method)

We want this because it's basically how LSP expects things to work, and VSCode, and it's how other languages work: they show up namespaces in the list of autocomplete results.

Where is the feature implemented? Can't be in nuclide LspLanguageService, because other LSP servers already do this work themselves. Can't be in hh_client because that has no knowledge of the text buffer. Therefore it must be implemented in hh_server.

Reviewed By: dabek

Differential Revision: D5366163

fbshipit-source-id: 6cd3bf9bf3aeb88bc2e54bd7267643c49950cee8