Allow autocomplete service to offer namespaces as completions
Summary:
This diff lets you autocomplete namespaces, e.g. `S| => Str`.

I had a tough decision on how to implement it. That's because the existing Hack indexes don't readily provide the namespace names. Here's an evaluation of the options.

Note: my overriding concern is that this new small feature should not introduce memory or perf regressions in the indexing service. I'm particularly mindful that there are often just a tiny handful of files that define namespaces, and thousands of files which don't define them.

Reviewed By: dabek

Differential Revision: D5366174

fbshipit-source-id: b0bdc6e777708e86e1f6e9b4990f85b67a90bd9e