Expose is_incomplete from hh_server's autocomplete
Summary:
The editor wants to know whether the autocomplete results from hh_server were exhaustive (from the current cursor position) or not. If they were exhaustive, then there's no need to re-compute autocomplete as the user types further characters -- we can do it purely by filtering the existing list.

Previously this was accomplished by hard-coding a magic number which was the limit of number of results. But I'd rather do this more explicitly.

Reviewed By: dabek

Differential Revision: D5368203

fbshipit-source-id: 119ff81e8eac392390b53bf144a7de29a24aebd1