fluid keystrokes for autocomplete namespaces
Summary:
I was exploring several options for how to make it more fluid when you type namespaces... (motivated by the fact that \ is an ugly character to type).

OPTION 1: `St|` autocomplete inserts the word `Str\|` with a trailing backslash. That's what the behavior was prior to this diff. Unfortunately when you accept this suggestion, although it automatically puts in the backslash, it doesn't automatically trigger autocomplete after that. In general editors don't automatically trigger a second autocomplete after the user accepts a prior one.

OPTION 2: `St|` autocomplete inserts the word `Str`. This leaves the user to type the backslash. This feels like it would be a fine solution, and in keeping with the rest of autocomplete.

OPTION 3: `St|` autocomplete inserts the word `Str\` but uses the magic of SNIPPETS so that the trailing backslash is a snippet. The user sees that this backslash is highlighted, and know they have to overtype the exact same symbol. It educates the user about what they have to type. Unfortunately, when you've got `Str\` then Nuclide highlights the entire word, and you can't see that the backslash is selected and ripe for overtype, so you end up typing additional characters like `compare` without overtyping the backslash first, and that doesn't work.

OPTION 4: `St|` autocomplete inserts the full `Str\` when you accept it with the key `\`, or it inserts just `Str` when you accept the suggestion via the Tab/Enter keys. This is how other editors work. This feels like the best option. But it will require serious tinkering with AutocompletePlus.

I've gone for option 2 for now. It's the least surprising.

Reviewed By: arxanas

Differential Revision: D5399141

fbshipit-source-id: 4c87348486bc60c9efa6e0c4f854fb921dfbb957