For loops, break, continue, and return; error reporting
Summary:
Mostly what it says on the tin. This also adds the very real possibility that
the activeBlock could be null (i.e. The last block had no continuation because
we made a non-structural jump) so a few checks for that need to be added.

`break` and `continue` now add the possiblity that the compiler encounters an
error that actually should be reported as a fatal error on the unit itself
instead of as an unexpected compiler failure, so this adds an exception class
for those errors and generates a unit that reports the error.

Also some glue added to actually get the filename into the output unit

Reviewed By: swtaarrs

Differential Revision: D5397795

fbshipit-source-id: 4d774c606f42dfd059e79f2c97198d68ed0d9b64