When out of retries, preserve line breaks from the original source
Summary:
The Line_splitter has a hard limit on the number of iterations it will attempt before giving up on solving a chunk group, since some chunk groups result in exponential-time formatting. When we reach this limit, we output the best solution found so far, which is generally [[ https://fb.prod.facebook.com/groups/1857379894541196/permalink/1923456957933489/ | extremely unappealing ]], with very long lines which the line splitter had not yet attempted to break. This behavior is inherited from dartfmt (see the [[ http://journal.stuffwithstuff.com/2015/09/08/the-hardest-program-ive-ever-written/#17 | "An escape hatch" ]] section).

After this change, instead of outputting the best solution so far as-is, we preserve the line breaks in the original source. This should be a much more acceptable behavior in the edge cases that encounter this problem.

Reviewed By: arxanas

Differential Revision: D5412096

fbshipit-source-id: 24c190c4a1be1c7dd0af182c3cfbdec0e261a0e1