Reduce max retries in line splitter
Summary:
Currently, the line splitter will repeat the process of dequeuing the best state so far, binding some of its unbound rules, and enqueuing the resulting states up to 2000 times.

I think this limit is higher than it needs to be. We can reduce it as low as 20 without failing any existing test cases. Now that we behave more reasonably when running out of retries (after D5412096), we can reduce this limit and ensure a more reasonable runtime. This change reduces the limit to 100.

Depends on D5412096

Reviewed By: arxanas

Differential Revision: D5412270

fbshipit-source-id: e45f67ed896590625a96b06823a5b038aac0c1bc