Add importer script for php7.x tests
Summary:
I've copied and gutted the existing import script for the Zend test suite.
There's a lot of functionality in the old script that we don't really need.
now, I'm having it just extract the release .tar and split the `phpt` files in two directories

I don't try to run the tests since the main goal is compatibility with the php7
emitter and it doesn't handle all that much right now. I'll add a blacklist and
whitelist of tests to run alongside the imported tests.

Reviewed By: swtaarrs

Differential Revision: D5381476

fbshipit-source-id: b61c5e13ddd4580443e0500cef4e8f514f161fec