Configure php7 tests
Summary:
Once the tests are imported, add a .ini for them and some lists of bad/passing tests

`bad-tests` crash or hang with HHVM's current emitter and runtime (many of
these are foreach-by-reference and bugs in the emitter with goto's)

`failing-tests` fail with the current emitter and runtime. Most of these are
relatively minor runtime differences I haven't sorted through yet. These are tests that fail when run with `--php7-runtime`

`passing-php7c-tests` pass with the current php7 emitter (with a patch I'm putting up next). These are tests that pass with `--php7``

Reviewed By: swtaarrs

Differential Revision: D5382508

fbshipit-source-id: 4dc492fd1edcc13f2e806d95ea06b8f0d2b89aab