Fix accounting for mallocDebt in MemoryManager::resetStatsImpl(false)
Summary:
resetStatsImpl(false) is called early in the request lifetime,
when MemoryManager has allocated some memory; we want to reset
external counters, but properly account for capacity/mallocDebt
going forward.

Currently we do this:
1. totalAlloc = mallocDebt
2. prevAllocated = *m_allocated // effectively zero out auxUsage
3. mallocDebt = 0

The next time we refreshStats, we'll end up with a correct value
for auxUsage and totalAlloc, but by doing it this way we've broken
a desireable invariant that capacity==(cumulative mallocDebt),
by zeroing mallocDebt too early.

This diff changes the accounting:
1. totalAlloc = 0
2. prevAllocated = *m_allocated - mallocDebt

This leaves the counters in the state they would have been if
the only external allocations were from MemoryManager. Specifically,
capacity==(cumulative mallocDebt) now and after the next refresh.

Reviewed By: alexeyt

Differential Revision: D5424283

fbshipit-source-id: 315140a8126fc2440f1c40a8a99bd2ac5546d78f