Fixing inconsistency of input to methodish_error helper functions
Summary: I remembered that one of the methodish_errors helper functions I'd written in the past took a different value as input than all the other helper functions, and I fixed that. Specifically, methodish_in_interface_has_body used to take the syntax of a node, while all the other helper functions directly took the node. I rewrote the offending function so that it also accepts the node now.

Reviewed By: ericlippert

Differential Revision: D5385174

fbshipit-source-id: dbe2941607f1c9acaf3545d5f0b853f1f9a860ad