Added support to specify files and directories
Summary:
This revision adds the support to specify specific files. If the file or directory doesn't exist it reports the number of missing files but continues executing.

I also considered just specifying the name of the file or directory and then let the script look inside the base directory but after talking to drphil we decided to leave like this since it's a common practice.

A following diff will add better output.

Reviewed By: hubyrod

Differential Revision: D5424129

fbshipit-source-id: df69f4674da14165fe3348ff66b74238bcbb2bed