Update Timeout.Alarm_timeout to use Timer
Summary:
`Timeout.Alarm_timeout` is the `Timeout` implementation that we use
everywhere except Windows. It relies on the alarm signal.

This fixes it up to use `Timer` instead of it's own implementation. This should
be better behaved for nested timeouts, as the outer won't pause while the inner
runs. The `Timeout` exceptions now indiciate exactly which timeout they
represent.

I've also slightly changed the API to avoid exposing the `Timeout` exceptions.
Instead, the `on_timeout` function is now required and its return value is used
in case of a timeout. The the calling code wants an exception, it can throw one
itself.

Reviewed By: alexchow

Differential Revision: D5229577

fbshipit-source-id: d46b3da866a46dc04840f9e912f2c4b56c51417d