Implement lowering of anonymous function coroutines
Summary:
We now lower anonymous function coroutines.

There is an error which will be corrected in the next diff: if there are two anonymous functions in the same function or method, they will both get the same name for their closure classes. I need to thread a unique number through the context, and use that to generate unique names.

Similarly, an anonymous function coroutine inside another coroutine will get the same name for its closure as the outer coroutine, which is wrong.

Reviewed By: michaeltingley

Differential Revision: D5373935

fbshipit-source-id: b4fc17418119376ce9f5dec4a1a945eeb878fb96