Implement closure class numbering for anon-method coroutines
Summary:
We were generating class names for closure classes based on the name of the enclosing class and method, but plainly this does not work for anonymous methods or lambdas, since there could be several of them in one method. (And the enclosing method could itself be a coroutine.)

We now generate unique names for all the closure classes associated with anonymous methods.

Next: lambdas.

Reviewed By: michaeltingley

Differential Revision: D5428860

fbshipit-source-id: 3757099d10066971cb7a685687daa49c114d3368