Fix a potential crash with profiled method dispatch
Summary:
If method dispatch profiling only sees a single method, but multiple
classes, it compiles in a fastpath which checks that the current class
has the right func at the expected offset. It also checks that the
class has at least that many methods - but the check was doing a 32
bit read of m_funcVecLen, which doesn't work in LowPtr builds.

Reviewed By: edwinsmith

Differential Revision: D5442987

fbshipit-source-id: e799d6b2bf5af8a7f8f509896b87f685a01305db