UPDATED: Improving method error recovery, making use of refactored parser context.
Summary:
This diff makes use of foundation laid in the previous five diffs, in order to do the same thing as D5384153 but without all the failing of tests.

In short, the issue was that items added onto the parser context never got popped off...so the parser thought tokens were expected later when they weren't, and it went off the rails as a result. The fix for this behavior happens in lines 560-564 of the declaration parser (and the equivalent lines in D5438161).

Below, I've reproduced the original message I wrote for the broken version of this diff.

----

Overview:

When inside a class, encountering a 'public,' 'private,' or 'protected' token inside an incomplete if_statement or methodish_declaration now bails out of the latter two scopes and continues parsing the-line-starting-with-the-visibility-modifier as a new methodish_declaration (rather than as a nested function, or as the body of an if statement). Testing shows that this greatly reduces the number of cascading errors.

Also, the parser context has been refactored to preserve a notion of "scope." Rather than storing context as a flat list of expected token_kinds, the token_kinds are bundled according to when in the context they became expected.

More detail:

This diff is another case-driven improvement to the full fidelity parser's error recovery. Specifically, it was motivated by the previously questionable recovery from the-programmer-hasn't-finished-typing-yet situations like those in test_nested_incomplete_if.php. It turns out that this file displayed two separate failures of parser error recovery, which I fixed.

Previous behavior:
* The parser would correctly identify the if statement as missing its right paren, but would then treat the declaration of function g() as part of the body of the if.
* Even without the if statement, error recovery was weird--see test_nested_function(.php/.old). A new line starting with a visibility modifier would be parsed as a function nested inside of method g().

Current behavior:
* The parser knows that 'public,' 'private', and 'protected' are valid tokens to encounter inside a class. So, when it's inside a class, but specifically in a scope where those tokens are unexpected--like in this incomplete if statement--it marks the rest of the if statement as 'missing' and bails out back up to the class level to continue parsing from there.
* When the parser is inside a method and encounters a 'public,' 'private,' or 'protected,' it marks the rest of the current method as 'missing' and starts parsing a new method at the visibility modifier.

Also, I refactored the parser context to include scope. I initially thought doing so would be necessary to solve the above issues, but that turned out to be untrue--my fixes don't make use of the scope feature. Still, I'm keeping the scope version around because it adds to the context extra information which might be useful later.

Notes:
- HHVM forbids functions-inside-functions from having visibility modifiers, but previous FFP behavior on test_nested_incomplete_if.php shows that the FFP let them slide. This diff has the side effect of fixing that, though\!
- PHP allows nested nominal functions, but Hack doesn't. TODO introduce an error for that later.
- The .exp files still contain multiple errors at the same location because I'm running --full-fidelity-errors-all. HOWEVER, the full fidelity unit tests compare the .exp files to the output of --full-fidelity-errors, so I've chosen to exclude test_nested_incomplete_if from the unit tests.
- The presence and absence of comments appears to affect the location of error reporting. For example, moving the line-38 comment in test_nested_function.php to line 37, after the brace, changes the reported missing-brace location from (37,32)-(38,1) to (38,1)-(38,1). TODO fix that.

Differential Revision: D5416550

fbshipit-source-id: 07f66af0b75c0440369fd636022e4437c1f44d20