Improving error recovery on extra tokens in function declarations
Summary:
So turns out this diff never actually broke anything; it was just stacked on top of one that did. I've made no changes to this code between this diff and the version of it that got reverted, but now the tests work! The original diff message is reproduced below.

------

**Overview:**

This diff is another case-based improvement to the full fidelity parser's error recovery. It implements, for method declarations, a mechanism for recovering from unknown extra tokens (which would be encountered often if code is being checked for syntax errors while the programmer is still typing). New and old behaviors are demonstrated in test_method_decl_extra_token.(php/exp/allexp/old).

**More Detail:**

This algorithm only works on unexpected words that come after an otherwise coherent modifiers list, and before a keyword from the set { 'function', 'coroutine', 'async' }. Any unexpected tokens fitting that description are marked missing and skipped.

...Unless the keyword in question has a leading newline\! Then we can reasonably infer that the programmer isn't accidentally breaking up a method declaration header, but that they have a function definition later and are in the process of defining a property. So, this diff also includes functionality for checking at parse-time whether or not a token has a leading newline.

//Notes://
- Unfortunately, since this error is generated at parse time, there is not yet a way to report it on the entire offending token (line 83 of full_fidelity_parser_helpers.ml says as much). TODO implement a mechanism for that next.
- @ Michael Tingley\! Just a heads up: in the process of implementing this, I made an improvement to the tests in hack/test/facebook/ffp/. Previously the master testing file would just let the user know that their test files produced syntax errors, but I modified it so that it now actually prints the errors out. I had to change some of your .exp files as a result of this. (At least, I think they're yours? I know you were working with shapes.)

Differential Revision: D5417338

fbshipit-source-id: 2b233dcb203a25051083e97da94fd8f16c4c2f50