Client Build should retry once on server hangup
Summary:
When the repo node is moved significantly, the Informant triggers
a typechecker restart (to prefer a saved state loading over a
incremental recheck).

However, when rebasing by a significant distance triggers a Hack
Build to regenerate files (via the Hg Merge Driver), the Informant directed
restart can cause that Hack Build to fail (due to the typechecker that is
doing the build being killed). Then the user has to manually retry the build
and the rebase with `hg resolve --all && hg rebase --continue`. The failure
happens in a race - when the hack client connects to the Hack
instance before the Informant has killed it first.

This adds a retry for when we see this condition.

Also removed is the need to wait for repe settling inside the typechecker.

This can be removed since it has been moved to the Monitor already in
D4848325

And it must be removed since the merge driver and the server will end up in
a cyclic deadlock - the merge driver cannot proceed until Hack
finishes init, and Hack init cannot proceed until the merge driver completes
(since the bad hack we use to "detect" whether or not the hg repo is mid-update
checks for the existence of an .updatestate sentinel file, which is present
the entire time the merge driver is running).

Reviewed By: jamesjwu

Differential Revision: D5442155

fbshipit-source-id: 002e5a3b81a128ec770b4a46c10b5ab335244890