hh_single_type_check --infer-return-types cleanup
Summary: Was previously filtering the list of types and functions to only get the ones in the the current file. However, the filter was actually trivially returning always true. Fixed it by limiting the check to only the file we care about in the first place. Also noticed that the position comparison that matches functions with their return types was reversed and fixed that.

Reviewed By: dlreeves

Differential Revision: D5448330

fbshipit-source-id: 3b527857dfd5657953dae0d470705954dd45363b