update handling of break/continue/return
Summary:
This diff changes the way how we emit break/continue/return statements.
Prior to that we used to emit these instructions as is and then rewrite code to properly handle nested loops or try/finally blocks. Now we maintain a stack of jump targets (loop, switch, try-body, finally) that are either targets of jumps or can alter the jump behavior (like jump from the try-body should go to the finally first) - this allows us to determine the target of the jump instruction at the point when we emit it without the need to rewrite them as the post processing step for loops. Also this diff adds a few more things to mimic HHVM:
- we use dedicated locals to store return value and label id
- we try to preserve the same id for the label:

```lang=php
while (1) {
  try {
    while (1) {
       try {
           if ($a) {
             break; // 0
           }
           else {
             continue 2; // 1
           }
       }
       finally {
         print "finally 2";
       }
    }
    if ($b) {
      continue; // 1
    }
  }
  finally {
    print "finally";
  }
}

```

Reviewed By: andrewjkennedy

Differential Revision: D5445865

fbshipit-source-id: 3b5abfc8adf8c16b385870fe2ed847390a26b171