Do not pass size to FreeList::push()
Summary:
In the past, we hoped to be able to kill the MM::initFree()
method, by having push() store a valid header (with size).
We've never managed to make that work without a regression,
and plumbing size through the code isn't worth it, just for
the sake of the remaining assertion.

Reviewed By: haoyangqian, alexeyt

Differential Revision: D5464306

fbshipit-source-id: b9a717250db347ced181ad2360516b147297dced