Testing harness for finding least upper bound of types
Summary:
This is a way for us to directly test the least upper bound algorithm
for types.

I chose to make it a part of hh_single_type_check because I wanted to sidestep
some messy startup (getting the typechecker/parser optons, file info,
file_name, etc.). However, I could try to put it in a stand alone file if
that's preferred.

Reviewed By: dlreeves

Differential Revision: D5455531

fbshipit-source-id: 47a16f5ce3bb66c7c44173cbce031ebc947baac5