Rehaul verify and hhbbc understanding of fault region edges and boundaries
Summary: The way that fault regions were generated for the php silence operator was also changed to include the Silence End in the fault region. Previously the Silence End was not included in the fault region, meaning that in cases with nested fault regions it was possible to create inconsistent states across exception edges. Additionally to this point, the verifier has been changed to check exception edges before each instruction in the fault region from it to its fault handler. The hhbbc emitter was also changed to not include inserted unconditional jumps in fault regions if these jumps leave the current fault region, as these conflict with the new way the verifier checks the exception edges.

Reviewed By: markw65

Differential Revision:
D5467406

Tags: hhvm

fbshipit-source-id: 997fa151666cb26e946e9752240c6642ebd737b5