Refactor coroutine context, state machine data, to contain outer variable information
Summary:
In this diff, a small refactoring in preparation for the next diff.

The problem is: if we have something like

    $x = whatever;
    $y = coroutine ($z) ==> { suspend m($x, $z); ... }

We need to ensure that (1) the initial values of $x and $z are passed in to the closure class constructor, and (2) ensure that $x and $z are preserved in the closure class across suspensions.  Right now we do (1) only for $z, not for $x; we do (2) for both.

Basically, we need to treat *closed-over outer variables of the lambda* as though they are parameters, because they kinda are.

This diff does not do so, but it prepares the way:

* The coroutine context now includes the parent chain used to compute the context.

* From the parent chain we can compute the outer variables of a lambda.

* These are then stored in the state machine data record.

In a subsequent diff we'll use the outer variables in the closure creation logic.

Reviewed By: vladima, michaeltingley

Differential Revision: D5465150

fbshipit-source-id: b2eaac1fa8e03ed2cbd2787af6c0e0943f181900