Update tc-print to work on Aarch64.
Summary:
These changes update the tc-print tool on Aarch64 hosts  so that it can print out the files generated with -vEval.DumpTC=true.  This reuses disassembler in the vixl module with a
slight change.  During Initial implementation of the Aarch64 host the capability of changing
the base address of the disassembler was removed from vixl.  This change adds back in part
of that.

This change also adds an option, -o, so that the binary opcodes are emitted in the report.
I find this useful in finding sequences to improve.  The default is off as before.

I do not have access to a powerpc64 host, but the disassembly code should be very similar.

These changes compile and run on both Aarch64 and x86_64 hosts.
Closes https://github.com/facebook/hhvm/pull/7703

Differential Revision: D4802760

Pulled By: mxw

fbshipit-source-id: 054297b417caff7bf110ec297f6935cdadf780dc