correction to ocaml Aarch64 code gen
Summary:
This corrects an issue in the ocaml compiler where it tries to make
a invalid immediate operand for a subtraction instruction.

This change was tested by rebuilding from scratch and running the
standard regression tests.

I'm not sure why I had to add the --ignore-whitespace option to the patch
command line.
Closes https://github.com/facebook/hhvm/pull/7918

Differential Revision: D5483019

Pulled By: mxw

fbshipit-source-id: 5a377d6826934c202ad1193cf09a1b59cc6cc2f6