fix potential crash in JSON_parser
Summary:
In the JSON_parser, there's a crash risk when a thread-local buffer allocation OOMs the request. The issue is that the thread-local sb_cap is updated to the new capacity even if the allocation fails. Thus, if the request OOMs, the next request on the thread will believe it still has a buffer of capacity sb_cap, when in fact it has no buffer.

To fix this, zero sb_cap when old buffer is freed and only set sb_cap to the new value when the allocation succeeds.

Reviewed By: ricklavoie

Differential Revision: D5473043

fbshipit-source-id: 9ebb3ede92cd1da0b1ebe698aab957906c861ca2