Better code gen for ClsCns
Summary:
If we know a base of the class, and that base has a declaration of the
constant, we know the constant's slot, and can load its value directly
from the class's constant table.

For dynamic constants, the loaded constant will be KindOfUninit, and
we'll fall back to the slow path.

Reviewed By: ricklavoie

Differential Revision: D5478555

fbshipit-source-id: 546753cbba401c6cd0157caf3db222eb8e2a6494