Add support for function definitions and calls
Summary:
Now that we have the machinery to deal with different flavors on the
eval stack, add functions. The only tricky bit here is the addition of
`Destination` instead of a bare flavor type since when passing args to a
function, some expressions (e.g. locals) have their own `FPass*` ops but all
FPass ops need to know which argument slot they're pushing to.

Reviewed By: swtaarrs

Differential Revision: D5438211

fbshipit-source-id: e222453ada52c131062c06edd10ba357b0b527a1