Fix graceful shutdown on signals
Summary:
If we get a SIGTERM/SIGHUP, we would tell the server to shutdown (if
it existed) and then reset the signal's default action and re-raise it.

If SIGTERM was set for graceful shutdown (which it is by default),
re-raising the signal would prevent that by killing the program immediately.

We should only re-raise the signal when there was no server running.

However, it turns out that KillOnSIGTERM=true causes hhvm to crash -
so to avoid that for now, we can re-raise the signal in that case to
preserve the existing behavior.

Reviewed By: jano

Differential Revision: D5488975

fbshipit-source-id: 1908139be7069f06aca80e9b7556676185b84c06