Consolidation of all return type inference code
Summary: Refactored all of the return type inference code so that there is no code copy/pasting and hh_single_type_check and hh_client work on the same underlying machinery. Also changed the way hh_single_type_check prints so that it tells you all the return types it found and then says try ___, which is the guessed upper bound type of that list.

Reviewed By: dlreeves

Differential Revision: D5475463

fbshipit-source-id: b22b39205d7b3f05016bd79274efbb2cc5cafe5b