Adding is_hack flag to distinguish between hack errors and php errors
Summary:
The FFP is currently used to parse php, but some of its error messages are triggered on behavior that's legal in php. To fix this, and to thereby allow the FFP to generate language-specific error messages, I added an is_hack flag to find_syntax_errors. I also used this flag to prevent error2038 from triggering on php code.

TODO: identify the other behaviors forbidden by the FFP that should be allowed in php, and allow them. (T20243683)

Reviewed By: ericlippert

Differential Revision: D5455527

fbshipit-source-id: ff31284487dc17792fcd2d02e4b44d1bbbbc40ee