Preserve line breaks in container literals
Summary:
If you write this in Javascript:

```
const foo = {
  bar: baz;
  qux: grault;
};
```

then Prettier will respect that you put the key-values on different lines, even though they could all fit on the same line. (This applies to even the case that there is only one key-value pair.)

This is desired behavior. Oftentimes we want to put members of a container literal on lines of their own for readability or emphasis. This commit changes it so that we look to see if the members of the container literal appear to be on different lines, and if so, preserves that formatting and keeps them on different lines.

Differential Revision: D5503181

fbshipit-source-id: f7d81e2ce909430161035524b0f442d6806b9fe2