Continue refactoring for lambda code generation
Summary:
I'm continuing to refactor the coroutine generation code in order to get lambda lowering right. It is tricky.

What we need to do is to determine which locals are (1) copied in and out in the body, (2) passed to the closure constructor, and (3) generated as properties in the closure body.

We don't do that quite correctly yet.  But the existing code was needlessly complicated; locals were passed around in a map when what was needed was a set, for instance.

I've simplified that code. Next, I'll improve the lambda analyzer so that it can partition the locals used in a body into (1) outer locals, (2) parameters, and (3) inner locals. Once we have that then the code to determine what we actually need will be much simplified.

Reviewed By: michaeltingley

Differential Revision: D5502710

fbshipit-source-id: 2556ff1715322089fd1aea10f6dd18a0efdfdb60