Bug fix in coroutine context generation
Summary:
Coroutine context generation was storing a parent list, but sometimes the parent list included the current node, and sometimes it did not.

As is often the case, when a data structure's contents are inconsistent, you get bugs.  Fixing this now prevents a bug from cropping up in a later refactoring which depends on the parent chain not including the current node.

Reviewed By: michaeltingley

Differential Revision: D5504467

fbshipit-source-id: 2331e2c02305de0dcd6826fb5b30758db576a8a5