Continue refactoring coroutines for lambda lowering
Summary: In order to correctly generate properties, etc, for coroutine closure classes we will need to classify all the locals used in a lambda into outer variables, inner variables, and formal parameters. In this refactoring, I create a tool in the lambda analyzer to do that, and fix up the closure data to be initialized from it.  Tweaks to the actual codegen will come in a subsequent diff.

Reviewed By: michaeltingley

Differential Revision: D5504520

fbshipit-source-id: 03bd16f1402b0441a9ed3575d328334360c64bb3