Verifier checks that Silence Ends have corresponding Starts
Summary: HHVM was crashing on bytecodes of the form SIlence x End, where x was a local variable id that had not previously been silenced with a Silence x Start. This adds a check to the verifier requiring that all codepaths to Silence x End contain a corresponding Silence x Start.

Reviewed By: markw65

Differential Revision: D5379702

fbshipit-source-id: e0a72aac79823c6b31e2e892b7abd6f0ac6970db