Set self in the environment correctly before NastCheck
Summary:
We never set the class self before doing a NastCheck. This made it so that the
this type always localized to TAny when checking expression dependent types,
hiding a huge array of errors in www, and even in our own unit tests! In the
deleted unit test, the code calls "this::TValue" even though no type
constant named TValue is even defined!

The test itself was probably based on code involving ElasticSearch in www,
which has the same problem: there are straight up references to non existent
type constants.

Reviewed By: dlreeves

Differential Revision: D5458671

fbshipit-source-id: e3ddb0f072e5ba523bc719e3487ff4ad1eda234b