Another lambda coroutine refactoring
Summary:
Two refactorings here. First, the "local_variables" field of the data record is renamed to what it will be used for: generating the closure class properties. Right now we get this slightly wrong; by naming it for what it means, we can more easily see what is wrong and that the fix is right.

Second, the set of synthetic locals that must be generated as properties was being stored in a map, unnecessarily. And also they were being generated in reverse order.  By turning it into a list instead of a map we solve both problems in one stroke.

Reviewed By: vladima

Differential Revision: D5512638

fbshipit-source-id: 3adb0bd7a019f2956a487308bfa603a8ea70c656