Adding a case of top level context-based error recovery to parser
Summary:
In this diff, I make progress towards comprehensively introducing context-based error recovery into the parser, from the top level down.

In what represents a bit of a shift of thinking, I've moved away from simply calling items on the parser context "expected tokens" and moved towards referring to them as "recovery tokens." This is because we don't necessarily want to use everything that's "expected" (allowed) in a particular scope as a sign that we should trigger recovery all the way back into that scope--perhaps certain tokens are legal in multiple scopes, for example? In that case, we wouldn't want to risk prematurely recovering too far up the stack.

Technically this risk is circumventable with perfect knowledge of what's legal when. However, for now I just want to focus on tokens that are reliable indicators of a context. In this particular example: if we wind up inside parse_classish_element, but we see Trait, Interface, or Class, something's gone wrong--none of these are legal here. But, they're legal at the top level, so we can recover to the top level.

This error recovery algorithm is tricky because it requires both a) deciding on a set of recovery tokens, b) deciding when to check for them. I placed this diff's check in parse_classish_element because as it turns out...classes inside of functions are legal in php!

(If you see what I'm getting at with 'recovery tokens', I'll go back and rename the other pushes to parser context to 'recovery_tokens'. I also considered 'anchor_tokens' but settled on this.)

Reviewed By: alexchow

Differential Revision: D5501483

fbshipit-source-id: b4f7fccfb1f713cf278f7f5235ecd2ba7c21ae01